Refactor facade
