+++ src/Exceptions/ProviderNotConfigured.php
@@ -8,6 +8,6 @@ class ProviderNotConfigured extends LogicException
 {
     public static function make(string $provider): self
     {
-        return new self('Provider "'.$provider.'" is not configured, please configure it in config/services.php.');
+        return new self('Provider "'.$provider.'" is not configured, please configure it in config/services.php and/or on your panel.');
     }
 }
