+++ src/FilamentSocialite.php
@@ -3,35 +3,14 @@
 namespace DutchCodingCompany\FilamentSocialite;
 
 use App\Models\User;
-use Closure;
 use DutchCodingCompany\FilamentSocialite\Exceptions\GuardNotStateful;
-use DutchCodingCompany\FilamentSocialite\Exceptions\ProviderNotConfigured;
-use DutchCodingCompany\FilamentSocialite\Models\Contracts\FilamentSocialiteUser as FilamentSocialiteUserContract;
 use Filament\Facades\Filament;
-use Filament\Panel;
-use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Contracts\Auth\Factory;
 use Illuminate\Contracts\Auth\StatefulGuard;
 use Illuminate\Contracts\Config\Repository;
-use Laravel\Socialite\Contracts\User as SocialiteUserContract;
 
 class FilamentSocialite
 {
-    /**
-     * @phpstan-var ?\Closure(string $provider, \Laravel\Socialite\Contracts\User $oauthUser, \DutchCodingCompany\FilamentSocialite\FilamentSocialite $socialite): ?(\Illuminate\Contracts\Auth\Authenticatable)
-     */
-    protected ?Closure $userResolver = null;
-
-    /**
-     * @phpstan-var ?\Closure(string $provider, \Laravel\Socialite\Contracts\User $oauthUser, self $socialite): \Illuminate\Contracts\Auth\Authenticatable
-     */
-    protected ?Closure $createUserCallback = null;
-
-    /**
-     * @phpstan-var ?\Closure(string $provider, \DutchCodingCompany\FilamentSocialite\Models\Contracts\FilamentSocialiteUser $socialiteUser): \Illuminate\Http\RedirectResponse
-     */
-    protected ?Closure $loginRedirectCallback = null;
-
     protected FilamentSocialitePlugin $plugin;
 
     public function __construct(
@@ -52,32 +31,6 @@ class FilamentSocialite
         return Filament::getCurrentPanel()->getPlugin('filament-socialite');
     }
 
-    public function isProviderConfigured(string $provider): bool
-    {
-        return $this->config->has('services.'.$provider);
-    }
-
-    /**
-     * @return class-string<\Illuminate\Contracts\Auth\Authenticatable>
-     */
-    public function getUserModelClass(): string
-    {
-        return $this->getPlugin()->getUserModelClass();
-    }
-
-    /**
-     * @return class-string<\DutchCodingCompany\FilamentSocialite\Models\Contracts\FilamentSocialiteUser>
-     */
-    public function getSocialiteUserModelClass(): string
-    {
-        return $this->getPlugin()->getSocialiteUserModelClass();
-    }
-
-    public function getSocialiteUserModel(): FilamentSocialiteUserContract
-    {
-        return new ($this->getSocialiteUserModelClass());
-    }
-
     public function getGuard(): StatefulGuard
     {
         $guard = $this->auth->guard(
@@ -89,39 +42,5 @@ class FilamentSocialite
         }
 
         throw GuardNotStateful::make($guardName);
-    }
-
-    /**
-     * @param \Closure(string $provider, \DutchCodingCompany\FilamentSocialite\Models\Contracts\FilamentSocialiteUser $socialiteUser): \Illuminate\Http\RedirectResponse $callback
-     */
-    public function setLoginRedirectCallback(Closure $callback): static
-    {
-        $this->loginRedirectCallback = $callback;
-
-        return $this;
-    }
-
-    /**
-     * @return \Closure(string $provider, \DutchCodingCompany\FilamentSocialite\Models\Contracts\FilamentSocialiteUser $socialiteUser): \Illuminate\Http\RedirectResponse
-     */
-    public function getLoginRedirectCallback(): Closure
-    {
-        return $this->loginRedirectCallback ?? function (string $provider, FilamentSocialiteUserContract $socialiteUser) {
-            if (($panel = Filament::getCurrentPanel())->hasTenancy()) {
-                $tenant = Filament::getUserDefaultTenant($socialiteUser->getUser());
-
-                if (is_null($tenant) && $tenantRegistrationUrl = $panel->getTenantRegistrationUrl()) {
-                    return redirect()->intended($tenantRegistrationUrl);
-                }
-
-                return redirect()->intended(
-                    $panel->getUrl($tenant)
-                );
-            }
-
-            return redirect()->intended(
-                route($this->getPlugin()->getDashboardRouteName())
-            );
-        };
     }
 }
